feat: 기계공학부와 그 외 학과의 크롤링 로직 통합
